+++ R/remove_neg.R
@@ -2,15 +2,27 @@
 
 negpar <- min(x)
 
-while (negpar < 0) {
+while (negpar <= 0) {
   #check for any negative parameters
-  omit <- which(x < 0)
+  omit <- which(x <= 0)
 
   #remove the column from the library that contains the identified data
-  if (length(which(x < 0)) > 0) {
+  if (length(which(x <= 0)) > 0) {
     lib$xrd <- lib$xrd[, -omit]
     x <- x[-omit]
   }
+
+  if (solver == "L-BFGS-B") {
+
+    cat("\n-Removing negative coefficients and reoptimising using L-BFGS-S constrained to a lower limit of zero...")
+    o <- stats::optim(par = x, .fullpat,
+                      method = solver, lower = 0, pure_patterns = lib$xrd,
+                      sample_pattern = smpl[, 2], obj = obj)
+    x <- o$par
+    #identify whether any parameters are negative for the next iteration
+    negpar <- min(x)
+
+  } else {
 
   cat("\n-Removing negative coefficients and reoptimising...")
   o <- stats::optim(par = x, .fullpat,
@@ -19,6 +31,9 @@ while (negpar < 0) {
   x <- o$par
   #identify whether any parameters are negative for the next iteration
   negpar <- min(x)
+
+  }
+
 }
 
 return(list("x" = x, "lib" = lib))